requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
Flask-SQLAlchemy==3.0.5
